bt3017_visual_learning_studio/requirements.txt
streamlit
numpy
matplotlib
scikit-learn
scipy
pandas
openai
python-dotenv